Safe contains usage
- Fixed, some string values not being escaped correctly for contains
expression.

+++ test/Mvc.Lookup.Tests/Objects/Models/NumericIdModel.cs
@@ -5,6 +5,6 @@ namespace NonFactors.Mvc.Lookup.Tests.Objects
     public class NumericIdModel
     {
         [LookupColumn]
-        public Decimal Id { get; set; }
+        public Int32 Id { get; set; }
     }
 }
